dashboard/requirements.txt
streamlit>=1.31
pandas>=2.1
pyyaml>=6.0
